+++ src/lib/Nodes/FlowNode.ts
@@ -1,5 +1,5 @@
 import { Assert } from '../Diagnostics/Assert.js';
-import { Fiber } from '../Graphs/Execution/Fiber.js';
+import { Fiber } from '../Execution/Fiber.js';
 import { Graph } from '../Graphs/Graph.js';
 import { Socket } from '../Sockets/Socket.js';
 import { Node } from './Node.js';
