+++ src/lib/Profiles/Core/Flow/Branch.ts
@@ -1,4 +1,4 @@
-import { Fiber } from '../../../Graphs/Execution/Fiber.js';
+import { Fiber } from '../../../Execution/Fiber.js';
 import { Graph } from '../../../Graphs/Graph.js';
 import { FlowNode } from '../../../Nodes/FlowNode.js';
 import { NodeDescription } from '../../../Nodes/Registry/NodeDescription.js';
