+++ src/lib/index.ts
@@ -16,8 +16,8 @@ export * from './Nodes/Templates/In1Out1FuncNode.js';
 export * from './Nodes/Templates/In0Out1FuncNode.js';
 
 // loading & execution
-export * from './Graphs/Execution/Engine.js';
-export * from './Graphs/Execution/traceToLogger.js';
+export * from './Execution/Engine.js';
+export * from './Execution/traceToLogger.js';
 export * from './Graphs/IO/readGraphFromJSON.js';
 export * from './Graphs/IO/writeGraphToJSON.js';
 export * from './Graphs/IO/writeNodeSpecsToJSON.js';
